+++ src/04_array/prime_factorization_hard.rb
